requirements.txt
certifi==2018.11.29
chardet==3.0.4
Deprecated==1.2.5
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
inflect==2.1.0
Jinja2==2.10
MarkupSafe==1.1.1
PyGithub==1.43.5
PyJWT==1.7.1
requests==2.21.0
smmap2==2.0.5
urllib3==1.24.1
wrapt==1.11.1
